visualizer/requirements.txt
matplotlib>=3.3.0
numpy>=1.19.0
mysql-connector-python>=8.0.0
